chore: Move language_selection_modal.dart to global widgets/
